Move ReviewControls out of Game.tsx.
